Refactored block sections storage
Instead of storing them in raw form, bow both raw and parsed versions
are stored. It turns out that the idea with raw only was impossible to
implement, though I didn't knew it at the time I started...

That was a hard update to implement this late.
Probably introduced lots of bugs.