Separated tag-string conversion
Now tag ID can be easily converted to string and vice versa.
This helps with debug of ScopeInfo issue. The value of ScopeInfo
should be automatically determined, though at the moment it does
not end up with the same values as Labview sets them.

Added warn message when the detection gives different value than
Labview uses.