Heap: Introduced ContentTagId
To identify sub-items assignment to enums, another parameter
was introduced - contentTagId. This has already turned out to
be bad decision - some nieche tags cannot be handled properly
due to this and their content need to be kept in raw (hex) form.

Still, this is better than nothing. It will have to be further
refactored, but works good enough and doesn't introduce regressions.

To fix this, classId and tagId need to be converted to enum right
away, while reading the input file. For unknown tags, we either need
a default class (that would make 'name' support easier), or keep
it as integer.