All: Implement LabVIEW attributes bug
The attribute of 7ffff is stored as 48-bit value instead of 16-bit.
LabVIEW can properly read files with that bug fixed, but we want
generated files to be identical to what LabVIEW makes.

This patch implements the bug to our parser, so that heaps generated
with it are binary identical to LabVIEW saved files.