DTHP: Write empty list like LV14
If there is no type to store in DTHP, original LabView does not
write anything. Then while reading, it still assumes the type is
there - but it reads padding, so it's not an issue.

This patch makes that work like in original - no connector index is now
written, but instead padding will be treated as 0. This allows
generating binary identical files to LV.