RSRC: Added option to print file map on extraction
During extraction, it is now possible to print a file map which clearly
shows what is stored at each offset of input file.

More details will likely be added in the future to the map.
It would also be better to store it as a file instead of printing
on terminal.