Classes: Preparing for auto setting version in variant
Storing a lot of the same version in input file is pointless.
If the version matches file version, it should be set to "auto".
But there is currently no call which goes through the tree after
all blocks are accessible, so no place to set the auto variable.

This introduces the calls required to set the auto variable,
and prepares outline of how it all should work.