HLPT: Introduced StupidZeroPadding tag
The way some HLPT blocks are written is just stupid. I assume these are
just remains of an old version of LV, but I still want to re-create them
in a bimary-identical way.

So there is now a tag which sets the size to which this stupid padding
needs to be extended. I really hope I won't have to spend any more time
with that thing.