HLPT: And here we go again
This block is completely irrelevant, yet it continues to force me to
work on it. Not because anything around it is meaningful, but because it
is so badly implemented that reproducing all the LV bugs requires
quite complex code.

Now this block has its own class, which supports reading its string
form. But if the damaged binary form is detected - parsing stops.