LIxx: Started refactoring
We currently support that block in a very limited way.
Now at least general structure if that block is properly
implemented.

Parts of the old implementation are not remade for the new way yet.

The new base class, CompleteBlock, can now be used for more blocks.
Using it will make implementations of the blocks which are completely
understood more robust, as it provides unified path of exceptions
support.