LIxx: Finished LinkObjHeapToRCFileLink implementation
That was a complex one. This LinkObj contains a list
of TD objects, which required reading similar to LVVariant,
but only the type. Separate set of functions was created for that.

Would be nice to now use these new functions also in LVVariant,
so that the code is not duplicated and easier to modify.