RSRC: Block size computation to support interleaved sections
Previously, the total size of a block was computed based on
start offset of first section, and a first section of a different
block above that offset. But sections of one block are not always
continous, sometimes they interleave - this caused incorrect
computations in some files, resulting in issues with parser.

Now each section size is computed individually, and then they
are summed up.